Tidy lint warnings in scripts/

+++ scripts/ablation_a2.py
@@ -33,10 +33,10 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 
-from pmhcpresent.io.pseudoseq import load_pseudosequences_json
-from pmhcpresent.train import PeptideMHCDataset, TrainConfig, train_model, evaluate
-from pmhcpresent.models.nn import PresentationNet, NetConfig
 from pmhcpresent.eval.splits import hamming_cluster
+from pmhcpresent.io.pseudoseq import load_pseudosequences_json
+from pmhcpresent.models.nn import NetConfig, PresentationNet
+from pmhcpresent.train import PeptideMHCDataset, TrainConfig, evaluate, train_model
 
 
 def make_split(df, allele_col, peptide_col, frac_val=0.2, seed=42):
@@ -44,7 +44,7 @@ def make_split(df, allele_col, peptide_col, frac_val=0.2, seed=42):
     clusters = hamming_cluster(df[peptide_col].tolist(), df[allele_col].tolist())
     rng = np.random.default_rng(seed)
     uniq = np.unique(clusters)
-    n_val = max(1, int(round(len(uniq) * frac_val)))
+    n_val = max(1, round(len(uniq) * frac_val))
     val_clusters = set(rng.choice(uniq, size=n_val, replace=False))
     is_val = np.array([c in val_clusters for c in clusters])
     return is_val
